Added more precision to sentiment values

+++ SocialCreditScoreBot2/ISentimentAnalyzer.cs
@@ -2,5 +2,5 @@ namespace SocialCreditScoreBot2;
 
 public interface ISentimentAnalyzer {
     public Task<bool> Init();
-    public Task<float> Analyse(string text);
+    public Task<double> Analyse(string text);
 }
